+++ backend/SEP/AgencyService/Models/ServiceOfferItem.cs
@@ -6,8 +6,8 @@ namespace AgencyService.Models
     {
         public EOfferName OfferName { get; set; }
         public bool IsAccepted { get; set; }
-        public int MonthlyPrice { get; set; }
-        public int YearlyPrice { get; set; }
+        public double MonthlyPrice { get; set; }
+        public double YearlyPrice { get; set; }
         public List<ServiceOffer>? ServiceOffers { get; set; }
     }
 }
